Try without instant navigation
Instant Navigation loads HTML in the background and patches it into the existing page. In theory this can be faster because it reduces time taken to process the search indexes.

However, it does seem to have a lot of bugs associated with it, and our site is not especially fast - so let's try without it.