Orange theme (#14)
Theme change to be visually distinct from the parent portal, to avoid creating navigation confusion. Otherwise the click-through appears to make the tab bar change silently, so (for example) it looks like the Developer Guides tab is moving around.

The colour change is kind of jarring - but this is probably the lesser of the two evils.

Hyperlinks darkened slightly for better contrast.